@@ -1,4 +1,4 @@
-class Patient < ActiveRecord::Base
+class Patient < ApplicationRecord
 	validates :cnic, presence: true, length: { is: 13 }
 
 end
