@@ -1,2 +1,3 @@
-class Request < ActiveRecord::Base
+class Request < ApplicationRecord
+	has_one :request_detail
 end
