@@ -1,2 +1,4 @@
-class RequestDetail < ActiveRecord::Base
+class RequestDetail < ApplicationRecord
+	belongs_to :request
+	validates :cnic, length: { is: 13 }
 end
