@@ -1,4 +1,3 @@
-# encoding: UTF-8
 # This file is auto-generated from the current state of the database. Instead
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
@@ -11,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20160728123050) do
+ActiveRecord::Schema.define(version: 20160811105355) do
 
   create_table "admins", force: :cascade do |t|
     t.string   "email"
@@ -28,8 +27,10 @@ ActiveRecord::Schema.define(version: 20160728123050) do
     t.string   "password_salt"
     t.string   "name"
     t.string   "belongs_to"
-    t.datetime "created_at",    null: false
-    t.datetime "updated_at",    null: false
+    t.datetime "created_at",                        null: false
+    t.datetime "updated_at",                        null: false
+    t.string   "authentication_token", default: ""
+    t.index ["authentication_token"], name: "index_ambulance_users_on_authentication_token", unique: true
   end
 
   create_table "beds", force: :cascade do |t|
@@ -38,9 +39,8 @@ ActiveRecord::Schema.define(version: 20160728123050) do
     t.datetime "created_at",                 null: false
     t.datetime "updated_at",                 null: false
     t.boolean  "vacant",      default: true
+    t.index ["hospital_id"], name: "index_beds_on_hospital_id"
   end
-
-  add_index "beds", ["hospital_id"], name: "index_beds_on_hospital_id"
 
   create_table "hospital_super_users", force: :cascade do |t|
     t.string   "user_name"
@@ -49,9 +49,8 @@ ActiveRecord::Schema.define(version: 20160728123050) do
     t.datetime "created_at",    null: false
     t.datetime "updated_at",    null: false
     t.integer  "hospital_id"
+    t.index ["hospital_id"], name: "index_hospital_super_users_on_hospital_id"
   end
-
-  add_index "hospital_super_users", ["hospital_id"], name: "index_hospital_super_users_on_hospital_id"
 
   create_table "hospitals", force: :cascade do |t|
     t.string   "name"
@@ -70,24 +69,25 @@ ActiveRecord::Schema.define(version: 20160728123050) do
     t.integer  "hospital_id"
     t.datetime "created_at",    null: false
     t.datetime "updated_at",    null: false
+    t.index ["hospital_id"], name: "index_managers_on_hospital_id"
   end
-
-  add_index "managers", ["hospital_id"], name: "index_managers_on_hospital_id"
 
   create_table "patients", force: :cascade do |t|
     t.integer  "cnic",       limit: 8, null: false
     t.datetime "created_at",           null: false
     t.datetime "updated_at",           null: false
+    t.index ["cnic"], name: "index_patients_on_cnic", unique: true
   end
-
-  add_index "patients", ["cnic"], name: "index_patients_on_cnic", unique: true
 
   create_table "request_details", force: :cascade do |t|
     t.string   "gender"
     t.string   "age"
     t.string   "blood_pressure"
-    t.datetime "created_at",     null: false
-    t.datetime "updated_at",     null: false
+    t.datetime "created_at",               null: false
+    t.datetime "updated_at",               null: false
+    t.integer  "request_id"
+    t.integer  "cnic",           limit: 8
+    t.index ["request_id"], name: "index_request_details_on_request_id"
   end
 
   create_table "requests", force: :cascade do |t|
@@ -99,10 +99,9 @@ ActiveRecord::Schema.define(version: 20160728123050) do
     t.boolean  "completed",         default: false
     t.boolean  "data_received",     default: false
     t.integer  "bed_id"
+    t.index ["ambulance_user_id"], name: "index_requests_on_ambulance_user_id"
+    t.index ["bed_id"], name: "index_requests_on_bed_id"
+    t.index ["hospital_id"], name: "index_requests_on_hospital_id"
   end
-
-  add_index "requests", ["ambulance_user_id"], name: "index_requests_on_ambulance_user_id"
-  add_index "requests", ["bed_id"], name: "index_requests_on_bed_id"
-  add_index "requests", ["hospital_id"], name: "index_requests_on_hospital_id"
 
 end
